requirements.txt
# Agentic AI Training Program - Full Dependencies
# Install with: pip install -r requirements.txt

# Core
python-dotenv==1.0.1
requests>=2.31.0

# OpenAI (primary track)
openai>=1.12.0

# Anthropic (Claude Code track)
anthropic>=0.21.0

# LangChain ecosystem (Weeks 6+)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.20

# LangSmith for debugging (optional)
langsmith>=0.0.75

# CrewAI (Weeks 4-5)
crewai>=0.28.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: visualization
# mermaid-py>=0.3.0  # For graph visualization
